Renaming and code optimization
