+++ web/src/lib/components/asset-viewer/detail-panel-star-rating.svelte
@@ -2,7 +2,7 @@
   import { handleError } from '$lib/utils/handle-error';
   import { updateAsset, type AssetResponseDto } from '@immich/sdk';
   import { t } from 'svelte-i18n';
-  import Star from '$lib/components/shared-components/settings/star-rating.svelte';
+  import StarRating from '$lib/components/shared-components/star-rating.svelte';
 
   export let asset: AssetResponseDto;
   export let isOwner: boolean;
@@ -14,27 +14,21 @@
 
   const handleChangeRating = async (clickedId: CustomEvent<number>) => {
     currentRating = clickedId.detail;
-    if (currentRating === asset.exifInfo?.rating) {
-      try {
-        await updateAsset({ id: asset.id, updateAssetDto: { rating: 0 } });
-      } catch (error) {
-        handleError(error, $t('errors.cant_apply_changes'));
-      }
-      return;
-    }
     try {
-      await updateAsset({ id: asset.id, updateAssetDto: { rating: currentRating } });
+      await updateAsset({
+        id: asset.id,
+        updateAssetDto: { rating: currentRating === asset.exifInfo?.rating ? 0 : currentRating },
+      });
     } catch (error) {
       handleError(error, $t('errors.cant_apply_changes'));
     }
-    rating = currentRating;
   };
 </script>
 
 <section class="relative flex px-4">
   {#each { length: countStars } as _, id}
     <span>
-      <Star {rating} {id} readOnly={!isOwner} on:click={handleChangeRating} />
+      <StarRating {rating} {id} readOnly={!isOwner} on:click={handleChangeRating} />
     </span>
   {/each}
 </section>
