+++ web/src/lib/components/shared-components/star-rating.svelte
