+++ web/src/lib/i18n/de.json
@@ -1022,6 +1022,7 @@
   "purchase_settings_server_activated": "Der Server Produktschlüssel wird durch den Administrator verwaltet",
   "range": "Reichweite",
   "rating": "Bewertung",
+  "rating_description": "Stellt die Exif-Bewertung im Informationsbereich dar",
   "raw": "RAW",
   "reaction_options": "Reaktionsmöglichkeiten",
   "read_changelog": "Changelog lesen",
@@ -1165,7 +1166,6 @@
   "show_password": "Passwort anzeigen",
   "show_person_options": "Personen-Optionen anzeigen",
   "show_progress_bar": "Fortschrittsbalken anzeigen",
-  "show_rating_details_panel": "Stellt die Exif-Bewertung im Informationsbereich dar",
   "show_search_options": "Suchoptionen anzeigen",
   "show_supporter_badge": "Unterstützer Abzeichen",
   "show_supporter_badge_description": "Zeige Unterstützer Abzeichen",
