+++ web/src/lib/i18n/en.json
@@ -957,6 +957,7 @@
   "purchase_server_title": "Server",
   "purchase_settings_server_activated": "The server product key is managed by the admin",
   "rating": "Star rating",
+  "rating_description": "Display the exif rating in the info panel",
   "reaction_options": "Reaction options",
   "read_changelog": "Read Changelog",
   "reassign": "Reassign",
@@ -1097,7 +1098,6 @@
   "show_password": "Show password",
   "show_person_options": "Show person options",
   "show_progress_bar": "Show Progress Bar",
-  "show_rating_details_panel": "Display the exif rating in the info panel",
   "show_search_options": "Show search options",
   "show_supporter_badge": "Supporter badge",
   "show_supporter_badge_description": "Show a supporter badge",
